Fixed downloadupdate dontdelete not respecting folder name

+++ src/updater/helpers/downloadupdate.js
@@ -4,7 +4,7 @@
  * Created Date: 09.07.2021 16:26:00
  * Author: 3urobeat
  * 
- * Last Modified: 26.02.2022 20:00:28
+ * Last Modified: 26.02.2022 20:05:26
  * Modified By: 3urobeat
  * 
  * Copyright (c) 2021 3urobeat <https://github.com/HerrEurobeat>
@@ -28,7 +28,7 @@ module.exports.downloadupdate = (releasemode, compatibilityfeaturedone, callback
     logger("", "Starting to download update...", true, false, logger.animation("loading"))
 
     const url = `https://github.com/HerrEurobeat/steam-comment-service-bot/archive/${releasemode}.zip`
-    const dontdelete = ["./.git", "./backup", "./src/data/cache.json", "./src/data/lastcomment.db", "./accounts.txt", "./customlang.json", "./logininfo.json", "./output.txt", "./proxies.txt", "./quotes.txt"]
+    const dontdelete = ["./src/data/cache.json", "./src/data/lastcomment.db", "./accounts.txt", "./customlang.json", "./logininfo.json", "./output.txt", "./proxies.txt", "./quotes.txt"]
 
     //Process dontdelete array in order to include parent folders of a dontdelete file in the array aswell
     dontdelete.forEach((e) => {
@@ -75,7 +75,7 @@ module.exports.downloadupdate = (releasemode, compatibilityfeaturedone, callback
         //Delete old files except files and folders in dontdelete
         logger("", `${logger.colors.fgyellow}Deleting old files...${logger.colors.reset}`, true, false, logger.animation("loading"))
         files.forEach((e, i) => {
-            if (fs.existsSync(e) && !dontdelete.includes(e) && !e.includes(`./steam-comment-service-bot-${releasemode}`) && !e.includes("./node_modules")) { //respect dontdelete, the fresh downloaded files and the node_modules folder
+            if (fs.existsSync(e) && !dontdelete.includes(e) && !e.includes(`./steam-comment-service-bot-${releasemode}`) && !e.includes("./node_modules") && !e.includes("./backup")) { //respect dontdelete, the fresh downloaded files and the node_modules folder
                 fs.rmSync(e, { recursive: true })
             }
